many improvements; fix route_via_t::replace_router
+use:noexcept
+use:final
~use size_t (instead of int) in src:crw:* and send_packet
~{apply_ping_cache_match -> ping_cache_match::apply}

+++ include/crw.h
@@ -2,15 +2,16 @@
 # define CRW_H 1
 # include <linux/if_tun.h>
 # include <netinet/in.h>
+# include <stddef.h>
 # ifdef __cplusplus
 extern "C" {
 # endif
   int tun_alloc(char *dev, const int flags);
-  int cread(const int fd, char *buf, const int n);
-  int cwrite(const int fd, const char *buf, const int n);
-  int crecvfrom(const int fd, char *buf, const int n, struct sockaddr_in *addr);
-  int csendto(const int fd, const char *buf, const int n, const struct sockaddr_in *addr);
-  int recv_n(const int fd, char *buf, const int n, struct sockaddr_in *addr);
+  int cread(const int fd, char *buf, const size_t n);
+  int cwrite(const int fd, const char *buf, const size_t n);
+  int crecvfrom(const int fd, char *buf, const size_t n, struct sockaddr_in *addr);
+  int csendto(const int fd, const char *buf, const size_t n, const struct sockaddr_in *addr);
+  int recv_n(const int fd, char *buf, const size_t n, struct sockaddr_in *addr);
 # ifdef __cplusplus
 }
 # endif
